git-svn-id: https://svn.lcsr.jhu.edu/cisst/branches/2011-12-15-fault@4589 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ cisst/cisstMultiTask/applications/cisstComponentGenerator/ccgParser.cpp
@@ -131,7 +131,7 @@ ccgMacro ccgParser::GetMacro(int startValue)
     for (ccgMacro macro = DECLARE_COMPONENT;
          macro < MAX_MACRO;
          macro = (ccgMacro) (macro + 1)) {
-        int temp;
+        size_t temp;
         if ((temp = Line.find(MacroNames[macro], startValue)) != -1) {
             Line = Line.substr(temp);
             return macro;
@@ -301,9 +301,9 @@ int ccgParser::ParseComponent(void)
 int ccgParser::ParseStateTableData(void)
 {
     std::string dtype, dname;
-    int startin = Line.find('(') + 1;
-    int endin = Line.find(',');
-    int namestart = Line.find(' ', endin) + 1;
+    size_t startin = Line.find('(') + 1;
+    size_t endin = Line.find(',');
+    size_t namestart = Line.find(' ', endin) + 1;
 
     dtype = Line.substr(startin, endin - startin);
     dname = Line.substr(namestart, Line.find(')') - namestart);
